uint16_t maps to i32 in djinni.

+++ src-gen/djinni/ClientImpl.hpp
@@ -27,11 +27,11 @@ public:
 
     static std::shared_ptr<ClientImpl> Open(ConfigurationImpl config, std::string client_id, std::shared_ptr<SubscribeCallbackImpl> subscribe_callback, std::shared_ptr<ReceiveCallbackImpl> receive_callback, SubscriptionStorage storage);
 
-    virtual PublishStatus Publish(int16_t namespace_id, std::string topic_name, RetentionBase retention, std::vector<uint8_t> data, std::experimental::optional<MsgIdImpl> message_id, std::shared_ptr<PublishCallbackImpl> publish_callback) = 0;
+    virtual PublishStatus Publish(int32_t namespace_id, std::string topic_name, RetentionBase retention, std::vector<uint8_t> data, std::experimental::optional<MsgIdImpl> message_id, std::shared_ptr<PublishCallbackImpl> publish_callback) = 0;
 
     virtual void ListenTopics(std::vector<SubscriptionRequestImpl> names) = 0;
 
-    virtual void Acknowledge(int16_t namespace_id, std::string topic_name, int64_t sequence_number) = 0;
+    virtual void Acknowledge(int32_t namespace_id, std::string topic_name, int64_t sequence_number) = 0;
 
     virtual void Close() = 0;
 };
